Add Contact as parent class to Phone, Email, Address
