+++ src/seedu/addressbook/data/person/Contact.java
@@ -1,0 +1,38 @@
+package seedu.addressbook.data.person;
+
+public class Contact {
+	public final String value;
+    protected boolean isPrivate;
+	
+    /**
+     * Construct new contact object
+     */
+    public Contact(String contact, boolean isPrivate) {
+        this.isPrivate = isPrivate;
+        contact = contact.trim();
+        this.value = contact;
+    }
+	
+
+    @Override
+    public String toString() {
+        return value;
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        return other == this // short circuit if same object
+                || (other instanceof Contact // instanceof handles nulls
+                && this.value.equals(((Contact) other).value)); // state check
+    }
+
+    @Override
+    public int hashCode() {
+        return value.hashCode();
+    }
+
+
+    public boolean isPrivate() {
+        return isPrivate;
+    }
+}
